REFACTOR : TestController dir refactor
created dir : testController
