+++ API_container/MadPtApi/src/main/java/com/example/MadPtApi/controller/testController/TEST_FoodController.java
@@ -1,4 +1,4 @@
-package com.example.MadPtApi.controller;
+package com.example.MadPtApi.controller.testController;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
